docker/laksh-back/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["ascorblack <a@scorblack.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
django-colorfield = "^0.11.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.40"
ujson = "^5.10.0"
django = "^5.2.5"
wagtail = "^7.1.1"
djangorestframework = "^3.16.1"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
